Create packStrips function
The packStrips function takes the first box and puts the scores from it
into the first strip.
Then, tries to find the next box with a score that is adjacent to the
score at the end of the strip, and that the boxWidths will not exceed
the total strip width.
If a box exists, puts scores from box into the first strip,
else, puts scores from box into next strip down.
This current method find the adjacent score that matches with the
SMALLEST scoreWidth, i.e. j = 0; j < numScores; ++j (unlike MTGMA).
Function then outputs all scores in all strips in order, and the total
width of all the strips.